requirements.txt
noobstuffs
psycopg2-binary
PyGithub
python-dotenv
python-telegram-bot
pytz
